+++ framework/contracts/account/manager/src/commands.rs
@@ -1,6 +1,8 @@
 use crate::{contract::ManagerResult, error::ManagerError, queries::query_module_version};
 use crate::{validation, versioning};
-use abstract_core::manager::state::{PENDING_GOVERNANCE, SUB_ACCOUNTS};
+use abstract_core::manager::state::{
+    PENDING_GOVERNANCE, REMOVE_ADAPTER_AUTHORIZED_CONTEXT, SUB_ACCOUNTS,
+};
 
 use abstract_core::adapter::{
     AdapterBaseMsg, AuthorizedAddressesResponse, BaseExecuteMsg, BaseQueryMsg,
@@ -46,6 +48,7 @@ use cw_storage_plus::Item;
 use semver::Version;
 
 pub const REGISTER_MODULES_DEPENDENCIES: u64 = 1;
+pub const HANDLE_ADAPTER_AUTHORIZED_REMOVE: u64 = 2;
 
 #[abstract_response(MANAGER)]
 pub struct ManagerResponse;
@@ -497,6 +500,12 @@ pub(crate) fn update_governance(deps: DepsMut, sender: &mut Addr) -> ManagerResu
     Ok(msgs)
 }
 
+// How many adapters removed, list of submessages
+pub(crate) struct RemoveAdapterAuthorized {
+    sent_count: u64,
+    sub_msgs: Vec<SubMsg>,
+}
+
 /// Migrate modules through address updates or contract migrations
 /// The dependency store is updated during migration
 /// A reply message is called after performing all the migrations which ensures version compatibility of the new state.
@@ -511,6 +520,11 @@ pub fn upgrade_modules(
     ensure!(!modules.is_empty(), ManagerError::NoUpdates {});
 
     let mut upgrade_msgs = vec![];
+
+    let mut remove_adapter_authorized = RemoveAdapterAuthorized {
+        sent_count: 0,
+        sub_msgs: vec![],
+    };
 
     let mut manager_migrate_info = None;
 
@@ -535,6 +549,7 @@ pub fn upgrade_modules(
                 module_info,
                 migrate_msg,
                 &mut upgrade_msgs,
+                &mut remove_adapter_authorized,
             )?;
         }
     }
@@ -542,7 +557,7 @@ pub fn upgrade_modules(
     // Upgrade the manager last
     if let Some((manager_info, manager_migrate_msg)) = manager_migrate_info {
         upgrade_msgs.push(self_upgrade_msg(
-            deps,
+            deps.branch(),
             &env.contract.address,
             manager_info,
             manager_migrate_msg.unwrap_or_default(),
@@ -554,19 +569,30 @@ pub fn upgrade_modules(
         &ExecuteMsg::Callback(CallbackMsg {}),
         vec![],
     )?;
+
+    // Save context
+    // TODO: remove after next patch
+    let RemoveAdapterAuthorized {
+        sent_count,
+        sub_msgs,
+    } = remove_adapter_authorized;
+    REMOVE_ADAPTER_AUTHORIZED_CONTEXT.save(deps.storage, &sent_count)?;
+
     Ok(ManagerResponse::new(
         "upgrade_modules",
         vec![("upgraded_modules", upgraded_module_ids.join(","))],
     )
+    .add_submessages(sub_msgs)
     .add_messages(upgrade_msgs)
     .add_message(callback_msg))
 }
 
-pub fn set_migrate_msgs_and_context(
+pub(crate) fn set_migrate_msgs_and_context(
     deps: DepsMut,
     module_info: ModuleInfo,
     migrate_msg: Option<Binary>,
     msgs: &mut Vec<CosmosMsg>,
+    remove_adapter_authorized: &mut RemoveAdapterAuthorized,
 ) -> Result<(), ManagerError> {
     let config = CONFIG.load(deps.storage)?;
     let version_control = VersionControlContract::new(config.version_control_address);
@@ -578,12 +604,21 @@ pub fn set_migrate_msgs_and_context(
 
     let migrate_msgs = match requested_module.module.reference {
         // upgrading an adapter is done by moving the authorized addresses to the new contract address and updating the permissions on the proxy.
-        ModuleReference::Adapter(new_adapter_addr) => handle_adapter_migration(
-            deps,
-            requested_module.module.info,
-            old_module_addr,
-            new_adapter_addr,
-        )?,
+        ModuleReference::Adapter(new_adapter_addr) => {
+            let (remove_authorized_submsg, msgs) = handle_adapter_migration(
+                deps,
+                requested_module.module.info,
+                old_module_addr,
+                new_adapter_addr,
+            )?;
+            remove_adapter_authorized
+                .sub_msgs
+                .extend(remove_authorized_submsg);
+            // we send 2 messages, one will succeed and other fail -
+            // so we count down to make sure we get expected amount of failed "updates"
+            remove_adapter_authorized.sent_count += 1;
+            msgs
+        }
         ModuleReference::App(code_id) => handle_app_migration(
             deps,
             migrate_msg,
@@ -611,7 +646,7 @@ fn handle_adapter_migration(
     module_info: ModuleInfo,
     old_adapter_addr: Addr,
     new_adapter_addr: Addr,
-) -> ManagerResult<Vec<CosmosMsg>> {
+) -> ManagerResult<([SubMsg; 2], Vec<CosmosMsg>)> {
     let module_id = module_info.id();
     versioning::assert_migrate_requirements(
         deps.as_ref(),
@@ -689,7 +724,7 @@ pub fn replace_adapter(
     deps: DepsMut,
     new_adapter_addr: Addr,
     old_adapter_addr: Addr,
-) -> Result<Vec<CosmosMsg>, ManagerError> {
+) -> Result<([SubMsg; 2], Vec<CosmosMsg>), ManagerError> {
     let mut msgs = vec![];
     // Makes sure we already have the adapter installed
     let proxy_addr = ACCOUNT_MODULES.load(deps.storage, PROXY)?;
@@ -705,27 +740,38 @@ pub fn replace_adapter(
         .into_iter()
         .map(|addr| addr.into_string())
         .collect();
-    // Remove authorized addresses from old
-    msgs.push(configure_adapter(
-        &old_adapter_addr,
-        BaseExecuteMsg {
-            msg: AdapterBaseMsg::UpdateAuthorizedAddresses {
+    // Remove authorized addresses
+    // If removing of authorized addresses on adapter failed - maybe it's old adapter
+    // and we need to retry it with old base message
+    let remove_authorized_submsg: SubMsg = SubMsg::reply_on_error(
+        configure_adapter(
+            &old_adapter_addr,
+            AdapterBaseMsg::UpdateAuthorizedAddresses {
                 to_add: vec![],
                 to_remove: authorized_to_migrate.clone(),
             },
-            proxy_address: None,
-        },
-    )?);
+        )?,
+        HANDLE_ADAPTER_AUTHORIZED_REMOVE,
+    );
+    // If removing of authorized addresses on adapter failed - maybe it's old adapter
+    // and we need to retry it with old base message
+    let remove_old_authorized_submsg: SubMsg = SubMsg::reply_on_error(
+        configure_old_adapter(
+            &old_adapter_addr,
+            AdapterBaseMsg::UpdateAuthorizedAddresses {
+                to_add: vec![],
+                to_remove: authorized_to_migrate.clone(),
+            },
+        )?,
+        HANDLE_ADAPTER_AUTHORIZED_REMOVE,
+    );
 
     // Add authorized addresses to new
     msgs.push(configure_adapter(
         &new_adapter_addr,
-        BaseExecuteMsg {
-            msg: AdapterBaseMsg::UpdateAuthorizedAddresses {
-                to_add: authorized_to_migrate,
-                to_remove: vec![],
-            },
-            proxy_address: None,
+        AdapterBaseMsg::UpdateAuthorizedAddresses {
+            to_add: authorized_to_migrate,
+            to_remove: vec![],
         },
     )?);
     // Remove adapter permissions from proxy
@@ -739,7 +785,10 @@ pub fn replace_adapter(
         vec![new_adapter_addr.into_string()],
     )?);
 
-    Ok(msgs)
+    Ok((
+        [remove_authorized_submsg, remove_old_authorized_submsg],
+        msgs,
+    ))
 }
 
 /// Update the Account information
@@ -926,9 +975,25 @@ fn remove_module_from_proxy(
 #[inline(always)]
 fn configure_adapter(
     adapter_address: impl Into<String>,
-    message: BaseExecuteMsg,
+    message: AdapterBaseMsg,
 ) -> StdResult<CosmosMsg> {
-    let adapter_msg: AdapterExecMsg<Empty> = message.into();
+    let adapter_msg: AdapterExecMsg = BaseExecuteMsg {
+        proxy_address: None,
+        msg: message,
+    }
+    .into();
+    Ok(wasm_execute(adapter_address, &adapter_msg, vec![])?.into())
+}
+
+// TODO: remove after patch
+#[inline(always)]
+fn configure_old_adapter(
+    adapter_address: impl Into<String>,
+    message: AdapterBaseMsg,
+) -> StdResult<CosmosMsg> {
+    type OldAdapterBaseExecuteMsg = abstract_core::base::ExecuteMsg<AdapterBaseMsg, Empty, Empty>;
+
+    let adapter_msg = OldAdapterBaseExecuteMsg::Base(message);
     Ok(wasm_execute(adapter_address, &adapter_msg, vec![])?.into())
 }
 
@@ -1022,6 +1087,16 @@ fn assert_admin_right(deps: Deps, sender: &Addr) -> ManagerResult<()> {
         }
         _ => Err(ownership_error),
     }
+}
+
+pub(crate) fn adapter_authorized_remove(deps: DepsMut, result: SubMsgResult) -> ManagerResult {
+    REMOVE_ADAPTER_AUTHORIZED_CONTEXT.update(deps.storage, |count| {
+        count
+            .checked_sub(1)
+            // If we got more errors than expected - return errors from now on
+            .ok_or(StdError::generic_err(result.unwrap_err()))
+    })?;
+    Ok(Response::new())
 }
 
 #[cfg(test)]