+++ framework/packages/abstract-sdk/src/cw_helpers/mod.rs
@@ -7,3 +7,4 @@ mod fees;
 pub use self::cosmwasm_std::*;
 pub use self::cw_storage_plus::*;
 pub use self::fees::*;
+pub use cw_clearable::*;