+++ modules/contracts/apps/subscription/src/handlers/execute.rs
@@ -5,6 +5,7 @@ use crate::state::{
     INCOME_TWA, SUBSCRIBERS, SUBSCRIPTION_CONFIG, SUBSCRIPTION_STATE,
 };
 use crate::SubscriptionError;
+use abstract_sdk::cw_helpers::Clearable;
 use abstract_sdk::{AbstractResponse, AccountAction, Execution, TransferInterface};
 use cosmwasm_std::{Addr, Decimal, Deps, DepsMut, Env, MessageInfo, Response, StdResult, Uint128};
 use cw_asset::{Asset, AssetInfoUnchecked};
@@ -41,7 +42,7 @@ pub fn execute_handler(
             payment_asset,
             subscription_cost_per_second,
             subscription_per_second_emissions,
-            unsubscription_hook_addr,
+            unsubscribe_hook_addr,
         } => update_subscription_config(
             deps,
             env,
@@ -50,7 +51,7 @@ pub fn execute_handler(
             payment_asset,
             subscription_cost_per_second,
             subscription_per_second_emissions,
-            unsubscription_hook_addr,
+            unsubscribe_hook_addr,
         ),
         SubscriptionExecuteMsg::RefreshTWA {} => {
             INCOME_TWA.try_update_value(&env, deps.storage)?;
@@ -199,7 +200,7 @@ pub fn unsubscribe(
         "unsubscribe",
     );
 
-    if let Some(hook) = subscription_config.unsubscription_hook_addr {
+    if let Some(hook) = subscription_config.unsubscribe_hook_addr {
         let msg = UnsubscribedHookMsg {
             unsubscribed: canceled_subs,
         }
@@ -297,7 +298,7 @@ pub fn update_subscription_config(
     payment_asset: Option<AssetInfoUnchecked>,
     subscription_cost_per_second: Option<Decimal>,
     subscription_per_second_emissions: Option<EmissionType<String>>,
-    unsubscription_hook_addr: Option<String>,
+    unsubscribe_hook_addr: Option<Clearable<String>>,
 ) -> SubscriptionResult {
     app.admin.assert_admin(deps.as_ref(), &info.sender)?;
 
@@ -317,8 +318,8 @@ pub fn update_subscription_config(
             subscription_per_second_emissions.check(deps.api)?;
     }
 
-    if let Some(human) = unsubscription_hook_addr {
-        config.unsubscription_hook_addr = Some(deps.api.addr_validate(&human)?);
+    if let Some(clearable_hook_addr) = unsubscribe_hook_addr {
+        config.unsubscribe_hook_addr = clearable_hook_addr.check(deps.api)?.into();
     }
 
     SUBSCRIPTION_CONFIG.save(deps.storage, &config)?;